Update README.md
Exploratory Data Analysis:
I always look for missing values and try to handle them. The dataset we are using is pre-cleaned so I think there is no missing value
Data Preprocessing:
Categorical features need to be converted to numbers so that they can be included in calculations done by a machine learning model
Model Creation and Evaluation:
We need to divide the dataset into training and test subsets so that we are able to measure the performance of our model on new, previously unseen examples.
Improving the Model:
GridSearchCV provides an easy way for parameter tuning. We can do cross-validation and try different parameters using GridSearchCV.